Fixed build.

+++ arrus/core/common/collections.h
@@ -241,7 +241,7 @@ std::vector<T> copyToVector(const Span<T> span) {
 
 template<typename T>
 bool areConsecutive(const std::vector<T> values) {
-    for(int i = 0; i < values.size(); ++i) {
+    for(size_t i = 0; i < values.size(); ++i) {
         if(values[i] != i) {
             return false;
         }
